Fix from Ignas Mikalajunas: I looked into your problem with timezones, and here is the best solution i could come up with (it is mostly right, i just don't know how (whether) you want to handle non existent ambiguous time). Pytz pays attention to dt when calculating dst, look at documentation of DstTzInfo.localize also http://pytz.sourceforge.net/.
Tuesday, July 25, 2006, 3:41:28 PM, Gary Poster wrote:
For the record, by the way, I think raising an error with an  
ambiguous time is ok for now; ideally we'd come up with some better  way to handle it in the UI (give the user options?  not sure) but  since the solution is not obvious, making it an error seems  
reasonable for now.